Add support for vector per target for hybrid
